Refactor: move bindings builder into its own module along with resolver and loader
